Fixed: MMU RAM having wrong size; Debug mX overflowing silently

+++ emu/cs/TinyProc/Memory/IMemoryDevice.cs
@@ -4,7 +4,7 @@ namespace TinyProc.Memory;
 
 public interface IMemoryDevice : IBusAttachable
 {
-    public uint Size { get; }
+    public ulong Size { get; }
     public uint ReadDirect(uint addr);
     public bool ReadEnable { get; set; }
 }
